Remove BrowserStorage
