+++ packages/common/random-access-storage/src/browser/idb-storage.ts
@@ -7,21 +7,26 @@ import { type RandomAccessStorage } from 'random-access-storage';
 
 import { invariant } from '@dxos/invariant';
 
-import { BrowserStorage } from './browser-storage';
-import { StorageType, getFullPath, wrapFile } from '../common';
+import { AbstractStorage, StorageType, getFullPath, wrapFile } from '../common';
 
 const DELIM = '\0';
 /**
  * Storage interface implementation for index DB.
  * https://github.com/random-access-storage/random-access-idb
  */
-export class IDbStorage extends BrowserStorage {
+export class IDbStorage extends AbstractStorage {
   public override type: StorageType = StorageType.IDB;
   private _db!: Promise<IDBDatabase>;
   private readonly _store = 'data';
   private _initialized = false;
+  protected readonly _fileStorage: (filename: string, opts?: {}) => RandomAccessStorage;
 
-  protected override _createFileStorage(path: string) {
+  constructor(path: string) {
+    super(path);
+    this._fileStorage = this._createFileStorage(path);
+  }
+
+  protected _createFileStorage(path: string): (filename: string, opts?: {}) => RandomAccessStorage {
     // Note: We use patched version of random-access-idb here that exposes the getdb method.
     const database = idb(path);
     let res: (db: IDBDatabase) => void;
@@ -31,6 +36,25 @@ export class IDbStorage extends BrowserStorage {
 
     database.getdb(res!);
     return database.create;
+  }
+
+  protected override async _destroy() {
+    // eslint-disable-next-line no-undef
+    return new Promise<void>((resolve, reject) => {
+      const request = indexedDB.deleteDatabase(this.path);
+      request.onsuccess = () => {
+        resolve();
+      };
+      request.onupgradeneeded = () => {
+        reject(new Error('Upgrade needed.'));
+      };
+      request.onblocked = () => {
+        reject(new Error('Blocked.'));
+      };
+      request.onerror = (err: any) => {
+        reject(err);
+      };
+    });
   }
 
   protected override _createFile(path: string, filename: string): RandomAccessStorage {
